Remove unnecessary cast

+++ NBitcoin.Altcoins/Argoneum.Internals/Lyra2Broken/Lyra2Broken.cs
@@ -150,7 +150,7 @@ namespace NBitcoin.Altcoins.ArgoneumInternals.Lyra2Broken
 
 		private void RunSetupPhase(int nBlocksInput, ulong nCols, ulong nRows)
 		{
-			for (i = 0; i < (int)nBlocksInput; i++)
+			for (i = 0; i < nBlocksInput; i++)
 			{
 				var arrayPassed = wholeMatrix.Skip((int)(i * (int)BLOCK_LEN_BLAKE2_SAFE_INT64 * (int)LyraConstants.UINT64_SIZE)).ToArray();
 				sponge.AbsorbBlockBlake2Safe(arrayPassed);
